Actualizo OFW (5.0.1 -> 5.0.3)

+++ ofw/core/OUpdate.php
@@ -5,8 +5,8 @@
 class OUpdate {
 	private $colors          = null;
 	private $base_dir        = null;
-	//private $repo_url        = 'https://raw.githubusercontent.com/igorosabel/Osumi-Framework/';
-	private $repo_url        = 'https://raw.githubusercontent.com/igorosabel/Osumi-Framework-Test/';
+	private $repo_url        = 'https://raw.githubusercontent.com/igorosabel/Osumi-Framework/';
+	//private $repo_url        = 'https://raw.githubusercontent.com/igorosabel/Osumi-Framework-Test/';
 	private $version_file    = null;
 	private $current_version = null;
 	private $repo_version    = null;
@@ -297,7 +297,9 @@ class OUpdate {
 				if (count($backups)>0) {
 					echo OTools::getMessage('TASK_UPDATE_DELETE_BACKUPS');
 					foreach ($backups as $backup) {
-						unlink($backup['backup']);
+						if (!is_null($backup['backup']) && file_exists($backup['backup'])){
+							unlink($backup['backup']);
+						}
 					}
 				}
 			}
